Create re-usable editable tag component
