+++ components/Act/ActDetailForm.js
@@ -2,7 +2,7 @@ import { Button, Divider, Form, Icon, Input, Tooltip } from 'antd'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import { FormattedMessage } from 'react-intl'
-import RichTextEditor from '../Editor/RichTextEditor'
+import RichTextEditor from '../FormInputComponents/RichTextEditor'
 import ImageUpload from '../UploadComponent/ImageUploadComponent'
 import { TextHeadingBold, TextP } from '../VTheme/VTheme'
 import OpDetailTagsEditable from '../Op/OpDetailTagsEditable'
